edge/jetson-segmentation/scripts/requirements.txt
numpy
pandas
matplotlib
glob2
scikit-image
tqdm
ipywidgets
ipython
seaborn
opencv-python
torch
torchvision
torchaudio
datasets
transformers
albumentations
wandb
torchmetrics
segmentation-models-pytorch
peft
Pillow
